Rename files and compose functions for better readability
